fbpreprocessor: only delete non-existent fw dumps
Before calling brillo::Delete(), check that the file does not exist.
Otherwise it adds an error message to the log.

BUG=b:291346820
TEST=FEATURES=test emerge-${BOARD} chromeos-base/fbpreprocessord
TEST=http://gpaste/6142262005202944, make sure that there are no errors
in the log.

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5008679
Tested-by: Nicolas Norvez <norvez@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>

+++ fbpreprocessor/firmware_dump.cc
@@ -7,6 +7,7 @@
 #include <ostream>
 
 #include <base/files/file_path.h>
+#include <base/files/file_util.h>
 #include <base/logging.h>
 #include <brillo/files/file_util.h>
 
@@ -24,11 +25,14 @@ base::FilePath FirmwareDump::BaseName() const {
 }
 
 bool FirmwareDump::Delete() const {
-  bool success = brillo::DeleteFile(dmp_file_);
-  if (!success) {
-    LOG(ERROR) << "Failed to delete firmware dump.";
+  if (base::PathExists(dmp_file_)) {
+    bool success = brillo::DeleteFile(dmp_file_);
+    if (!success) {
+      LOG(ERROR) << "Failed to delete firmware dump.";
+    }
+    return success;
   }
-  return success;
+  return true;
 }
 
 std::ostream& operator<<(std::ostream& os, const FirmwareDump& dump) {
